+++ test/csv_generator_test.exs
@@ -11,6 +11,7 @@ defmodule CsvGeneratorTest do
     column :calculated, :float, digits: 1, source: :points, with: &calc/1
     hardcoded :date, "today", ~D[2020-03-29]
     column :hour, :time
+    column :q, :float, with: fn x -> if x == nil, do: 0, else: x end, digits: 1
 
     def calc(pt) do
       1000 / pt
@@ -20,12 +21,12 @@ defmodule CsvGeneratorTest do
   test "rendering" do
     assert MyCSV.render([
              %{name: "Herman", points: 122, hour: 7200},
-             %{name: "Jose", points: 902, hour: 34231}
+             %{name: "Jose", points: 902, hour: 34231, q: 0.192}
            ]) ==
              """
-             "player","points","calculated","today","hour"
-             "Herman",122,8.2,2020-03-29,02:00
-             "Jose",902,1.1,2020-03-29,09:30\
+             "player","points","calculated","today","hour","q"
+             "Herman",122,8.2,2020-03-29,02:00,0.0
+             "Jose",902,1.1,2020-03-29,09:30,0.2\
              """
   end
 end